### 2019-05-30 #7 @marskar
Nbless: Software Submission for Review
Submitting Author: Martin Skarzynski (@marskar)
Package Name: Nbless
One-Line Description of Package: Construct, deconstruct, convert, execute, & prepare slides from Jupyter notebooks
Repository Link:  https://github.com/marskar/nbless/
Version submitted: 0.2.26
Version accepted: TBD

---

    The Nbless Python package allows you to
        - (de)construct,
        - convert,
        - execute, and
        - prepare slides from
    Jupyter notebooks.

    The Nbless Python package consists of 6 Python functions and shell commands:
        - `nbconv`, which converts a notebook into various formats.
        - `nbdeck`, which prepares a notebook to be viewed as or converted into slides.
        - `nbexec`, which runs a notebook from top to bottom and saves an executed version.
        - `nbless`, which calls nbuild and nbexec to create and execute a notebook.
        - `nbraze`, which extracts code and markdown files from a notebook.
        - `nbuild`, which creates a notebook from code and markdown files.

    For more on Nbless, please take a look at the [official documentation](https://marskar.github.io/nbless/readme.html).

- Please indicate which [category or categories](https://www.pyopensci.org/dev_guide/peer_review/aims_scope.html) this package falls under:

\* Please fill out a pre-submission inquiry before submitting a data visualization package. For more info, see [this section](https://www.pyopensci.org/dev_guide/peer_review/aims_scope.html#notes-on-categories) of our guidebook.

    Jupyter notebooks are important for reproducibility, because they facilitate literate programming, the combination of code and text.
    Literate programming is advantageous, because it allows descriptions of methods and results to be adjacent to the code used to obtain the results.
    I frequently use Jupyter notebooks in my work and in the Python classes I teach, but I often wish I could share small, simple files instead of large Jupyter notebooks.
    With Nbless, anyone can create Jupyter notebooks from simple source files or extract code and markdown files from Jupyter notebooks.
    In this way, Nbless supports round-trip conversion between source files and Jupyter notebooks. In other words, code and markdown files can be extracted from a Jupyter notebook _AND_ then used to recreate the original Jupyter notebook.
    This can promote the modularity principle without sacrificing the benefits of literate programming or the convenience of a single file with all code, text, and output for the project.
    Code and markdown files can be easily be accessed with any text editor or even opened in a browser without an internet connection. Keeping source files makes version control simpler than trying to track changes made to Jupyter notebooks (see [`nbdime`](https://github.com/jupyter/nbdime#nbdime-jupyter-notebook-diff-and-merge-tools) for another solution). Code files (`.py`, `.R`, etc.) are amenable to testing and linting, while notebooks are difficult to lint/test (see [`autopep8`](https://github.com/kenkoooo/jupyter-autopep8#jupyter-autopep8-pep8-formatter-for-jupyter-notebook) and [`nbval`](https://github.com/computationalmodelling/nbval#pytest-plugin-for-validating-jupyter-notebooks)).

    The target audience is anyone who uses Jupyter notebooks. This package may help teachers and researchers to improve the efficiency of their Jupyter notebook workflows. One of the major features of Nbless is that it can facilitate the creation of slides from Jupyter notebooks. Jupyter slides can be [converted in `revealjs` HTML files](https://github.com/marskar/slides#create-html-slides-from-ipynb-using-nbconvert-from-the-command-line) or viewed using [`nbviewer` or RISE/Binder](https://github.com/marskar/slides#create-html-slides-from-ipynb-using-nbconvert-from-the-command-line).
    I can imagine teachers and researchers writing markdown and code files and then using `nbless` to creating Jupyter notebooks to share with their students and colleagues.

    Related packages include:

    - [`pandoc`](https://pandoc.org/MANUAL.html#creating-jupyter-notebooks-with-pandoc)
    - [`jupytext`](https://github.com/mwouts/jupytext)
    - [`notedown`](https://github.com/aaren/notedown)

    These packages can all convert Jupyter notebooks to other formats: markdown files (all three) or Python scripts (`jupytext`). Nbless wraps `nbconvert` to convert between file types, but it can do something all of the aforementioned packages cannot. Nbless can take a more modular approach to file conversion by extracting the contents of each notebook cell into a separate file (cell -> file) or using a source file to create each notebook cell (file -> cell). Looking beyond simple file conversion, Nbless includes a tool for making slides from notebooks (by setting `slide_type` in notebook metadata).

-   If you made a pre-submission enquiry, please paste the link to the corresponding issue, forum post, or other discussion, or @tag the editor you contacted:

    The pre-submission inquiry link is https://github.com/pyOpenSci/software-review/issues/4

    I will close #4 as soon as I am done with #7.

For details about the pyOpenSci packaging requirements, see our [packaging guide](https://www.pyopensci.org/dev_guide/packaging/packaging_guide.html). Confirm each of the following by checking the box.  This package:

- [x] has an [OSI approved license](https://opensource.org/licenses)

- [ ] The package has an **obvious research application** according to JOSS's definition in their [submission requirements](https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements). Be aware that completing the pyOpenSci review process **does not** guarantee acceptance to JOSS. Be sure to read their submission requirements (linked above) if you are interested in submitting to JOSS.
- [ ] The package is not a "minor utility" as defined by JOSS's [submission requirements](https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements): "Minor ‘utility’ packages, including ‘thin’ API clients, are not acceptable." pyOpenSci welcomes these packages under "Data Retrieval", but JOSS has slightly different criteria.

## Are you OK with Reviewers Submitting Issues to your Repo Directly?

- [x] I agree to abide by [pyOpenSci's Code of Conduct](https://www.pyopensci.org/dev_guide/peer_review/coc.html) during the review process and in maintaining my package should it be accepted.

**P.S.** *Have feedback/comments about our review process? Leave a comment [here](https://github.com/pyOpenSci/governance/issues/8)*
