### 2019-05-30 #7 @lwasser
## Editor checks:

- [x] **Automated tests:** Package has a testing suite and is tested via Travis-CI or another CI service. TRAVIS
- [x ] **License:** The package has an OSI accepted license MIT

---

## Editor comments
Thank you for this submission @marskar  @ocefpaf and @leouieda will begin their review
Note: Martin is open to reviewers submitting issues to the repo directly. So please as you create your review, link to those issues if you create them
---

Reviewers: @ocefpaf @leouieda
Due date: 24 June 2019

@ocefpaf i'm suggesting a review deadline of June 24. It sounded like you wanted to do reviews one at a time. So @ocefpaf will submit his first and @leouieda ... will the two of you please coordinate on that? this is a new process so let's figure out what works best. also new is submitting issues directly rather than responding to a lot of text. A link to the reviewer template is below. I was thinking about adding readme checks to the reviewer template. please let me know what you think about that. i think it might make things easier!

[Editor and review templates can be found here](https://www.pyopensci.org/dev_guide/appendices/templates.html)
