### 2019-06-02 py4ds/nbless#12 @ocefpaf
Two different versions are defined
The installed PKG_INFO reports `0.2.26` but the package `__version__` reports `0.2.24`. Looks like the version is defined both in `setup.py` and `__init__.py`. My suggestion is to user a pattern to get the version for a single place, or use something like `versioneer` or `setuptools` `pkg_resources`.

xref.: https://github.com/pyOpenSci/software-review/issues/7
