### 2019-06-02 py4ds/nbless#13 @ocefpaf
nbexec behavior is different from the docs
When using `nbexec` I get an output in the `stdout` but the docs mentioned I should get a suffixed file. Here is the docs text:

> The nbexec command creates a copy of the input notebook, runs it from top to bottom and saves it. If an out_file name is not provided, the new filename will be the original filename with _executed.ipynb appended to it.

xref.: https://github.com/pyOpenSci/software-review/issues/7
