### 2019-06-02 py4ds/nbless#14 @ocefpaf
The docs mention two ways for converting to HTML but only one works
I was able to successfully use:

```
nbconv notebook.ipynb -e html
```

but

```
nbconv notebook.ipynb -o report.html
```

 created a text file and not valid HTML.

xref.: https://github.com/pyOpenSci/software-review/issues/7
