### 2019-06-02 py4ds/nbless#15 @ocefpaf
Confusing doc section "basic-usage-python-environment"
The section https://marskar.github.io/nbless/readme.html#basic-usage-python-environment is quite confusing. Looks like a work-flow example and should probably renamed in the docs, maybe replaced by a link to the file?

xref.: https://github.com/pyOpenSci/software-review/issues/7
