### 2019-06-02 py4ds/nbless#16 @ocefpaf
README is a copy of the docs and can be simplified.
This is my opinion and not really a PyOpenSci recommendation but README should be straight to the point, with some examples and a quick way to get the package installed.

xref.: https://github.com/pyOpenSci/software-review/issues/7
