### 2019-06-02 py4ds/nbless#17 @ocefpaf
Tests cover only 2 out 6 main CLI
xref.: https://github.com/pyOpenSci/software-review/issues/7
