### 2019-06-02 #7 @ocefpaf
@lwasser, @leouieda and @marskar here is my review.
@marskar please feel free to ignore the checked items when copying this to answer and use only the unchecked ones. Also, every actionable item has an issue to the package repository linked, so feel free to answer there first and then summarize/cross-reference it here.

- [X] **Function Documentation:** for all user-facing functions
- [X] **Examples** for all user-facing functions
- [ ] **Community guidelines** including contribution guidelines in the README or CONTRIBUTING, and DESCRIPTION with `URL`, `BugReports` and `Maintainer`.

My suggestion is to reduce the overlap of the README with the documentation and make it more to the point: install and use examples (https://github.com/marskar/nbless/issues/16). There is one section in the docs that is quite confusioning and should probably be just an example of a workflow instead, with a link to the file (https://github.com/marskar/nbless/issues/15). The package reports two different version numbers (https://github.com/marskar/nbless/issues/12).

I found two minor problems when trying to run the CLI as the docs recommend:

- https://github.com/marskar/nbless/issues/14
- https://github.com/marskar/nbless/issues/13

Test covers only 2 out of 6 main functions (https://github.com/marskar/nbless/issues/17).

- [X] **Continuous Integration:** Has continuous integration, such as Travis CI, AppVeyor, CircleCI, and/or others.

---

The package is a good fit for PyOpenSci. The documentation could use some polishing, the CLI examples are good but the rest is a bit terse to read. I believe that one of the functionality issues I found may be due to documentation problems and not really a bug (https://github.com/marskar/nbless/issues/13).

The repository does not have a community guidelines. (Note to PyOpenSci: do we have examples of those to point people to?)

Both development and standard installations were easy and all the tests passed. My doubt is regarding the `pandoc` dependency declared on the CI. Pandoc is optional in `nbconvert` but I know that some results may be different if you have it present. Does `nbless` expect `pandoc` for some features? if so are those tested properly? The tests passes without it BTW but I found [this comment](https://github,.com/marskar/nbless/blob/master/tests/test_nbless.py#L43) and I could not find a test against the expected output.
