### 2019-06-02 #7 @marskar
Thanks for the very helpful review @ocefpaf!
I plan to work through each of the 6 issues before the end of the week.

In the meantime, the answer to your question regarding pandoc is that the `nbconv` function/command (a wrapper for `jupyter nbconvert`) requires `pandoc` for [conversion to all formats other than HTML](https://nbconvert.readthedocs.io/en/latest/install.html#installing-pandoc). Converting to PDF also requires LaTeX. I will make sure to include this in the docs.
The `nbconv` tests that you linked to are commented out, because I had trouble satisfying the pandoc dependency in my Travis build. I will try to figure out how to get `pandoc` and Travis to cooperate.
