### 2019-06-03 97b6229
add pypandoc to requirements to install pandoc on windows and mac
