+++ requirements.txt
@@ -1,7 +1,14 @@
 -e .
 jupyter==1.0.0
 click==7.0
-pytest==4.6.0
+pytest==4.6.1
 pytest-mypy==0.3.2
 matplotlib==3.1.0
 numpy==1.16.4
+pypandoc==1.4
+bumpversion==0.5.3
+wheel==0.33.4
+tox==3.12.1
+coverage==4.5.3
+Sphinx==2.1.0
+twine==1.13.0
