### 2019-06-04 7b717e8
Update pytest from 4.6.1 to 4.6.2
