+++ requirements.txt
@@ -1,7 +1,7 @@
 -e .
 jupyter==1.0.0
 click==7.0
-pytest==4.6.1
+pytest==4.6.2
 pytest-mypy==0.3.2
 matplotlib==3.1.0
 numpy==1.16.4
