+++ requirements.txt
@@ -10,5 +10,5 @@ bumpversion==0.5.3
 wheel==0.33.4
 tox==3.12.1
 coverage==4.5.3
-Sphinx==2.1.0
+Sphinx==2.1.1
 twine==1.13.0
