### 2019-06-13 526dc06
add nbexec test and enable nbconv to infer exporter type
