+++ src/nbless/nbexec.py
@@ -6,7 +6,7 @@ from nbconvert.preprocessors import ExecutePreprocessor
 from nbformat.notebooknode import NotebookNode
 
 
-def nbexec(in_file: str, kernel: str = "python3") -> Tuple[str, NotebookNode]:
+def nbexec(in_file: str, kernel: str = "python3") -> NotebookNode:
     """Create an executed notebook without modifying the input notebook.
 
     :param in_file: The name of the Jupyter notebook file to be executed.
