+++ tests/test_nbless.py
@@ -2,9 +2,9 @@ from pathlib import Path
 from typing import List
 
 import nbformat
-# import pytest
+import pytest
 
-from nbless import nbless, nbuild#, nbconv
+from nbless import nbless, nbuild, nbexec, nbconv
 
 
 def make_tempfiles(tmp_path: Path) -> List[str]:
@@ -40,19 +40,26 @@ def test_nbless(tmp_path: Path) -> None:
     for tempfile in make_tempfiles(tmp_path):
         assert nbless([tempfile]).cells[0].source == Path(tempfile).read_text()
 
-# Can't get Pandoc to work with Travis
 
-# @pytest.mark.parametrize('not_exporters', ['htm', 'ipython', 'markup'])
-# def test_raises(not_exporters, tmp_path: Path) -> None:
-#     """Make sure a ValueError is raised if nbconv() gets a bad exporter."""
-#     nb = make_temp_notebook(tmp_path)
-#     with pytest.raises(ValueError):
-#         nbconv(in_file=nb, exporter=not_exporters)
+def test_nbexec(tmp_path: Path) -> None:
+    """Run nbexec() to execute a temporary notebook file."""
+    for cell in nbexec(make_temp_notebook(tmp_path)).cells:
+        if cell.cell_type == 'code':
+            assert cell.execution_count
+            for output in cell.outputs:
+                assert output
+
+@pytest.mark.parametrize('not_exporters', ['htm', 'ipython', 'markup'])
+def test_raises(not_exporters, tmp_path: Path) -> None:
+    """Make sure a ValueError is raised if nbconv() gets a bad exporter."""
+    nb = make_temp_notebook(tmp_path)
+    with pytest.raises(ValueError):
+        nbconv(in_file=nb, exporter=not_exporters)
 
 
-# @pytest.mark.parametrize('exporters', ['html', 'asciidoc', 'rst'])
-# def test_nbconv(exporters, tmp_path: Path) -> None:
-#     """Convert ``tempfiles`` with each exporter in ``exporters``."""
-#     nb = make_temp_notebook(tmp_path)
-#     assert nbconv(in_file=nb,
-#                   exporter=exporters)[0].endswith("." + exporters)
+@pytest.mark.parametrize('exporters', ['html', 'asciidoc', 'rst'])
+def test_nbconv(exporters, tmp_path: Path) -> None:
+    """Convert ``tempfiles`` with each exporter in ``exporters``."""
+    nb = make_temp_notebook(tmp_path)
+    assert nbconv(in_file=nb,
+                  exporter=exporters)[0].endswith("." + exporters)
