### 2019-06-13 marskar/nbless#14 @marskar
I fixed this in the code with this commit: 526dc06
I will not close this issue until I write [click tests](http://click.palletsprojects.com/en/5.x/testing/) to make sure the CLI behaves as expected.
