### 2019-06-13 marskar/nbless#17 @marskar
I added a test for nbexec with this commit: 526dc06
I will not close this issue until I write click tests to make sure the CLI behaves as expected.
