### 2019-06-13 d8af3e6
fix mypy error in nbconv, add to tests, and try to set travis home for pandoc
