### 2019-06-13 ebf215e
try to fix travis pandoc
