+++ .travis.yml
@@ -4,8 +4,8 @@ python:
 - 3.6
 - 3.7
 install:
-  - export HOME=/home/travis
   - pip install -U tox-travis pytest-mypy jupyter
+  - sudo apt-get install pandoc
 script: tox
 deploy:
   provider: pypi
@@ -17,6 +17,3 @@ deploy:
     tags: true
     repo: marskar/nbless
     python: 3.7
-addons:
-  apt_packages:
-    - pandoc
