+++ .circleci/config.yml
@@ -15,12 +15,12 @@ jobs:
           name: install dependencies
           command: |
             python3 -m venv .venv
-            . venv/bin/activate
+            . .venv/bin/activate
             pip install -r requirements.txt
       # Step 3: run linter and tests
       - run:
           name: run tests
           command: |
-            . venv/bin/activate
+            . .venv/bin/activate
             flake8 --exclude=.venv* --statistics
             pytest -v
