+++ requirements.txt
@@ -2,6 +2,7 @@
 jupyter==1.0.0
 click==7.0
 pytest==4.6.1
+black==19.3b0
 pytest-mypy==0.3.2
 matplotlib==3.1.0
 numpy==1.16.4
