+++ .circleci/config.yml
@@ -22,5 +22,5 @@ jobs:
           name: run tests
           command: |
             . .venv/bin/activate
-            flake8 --exclude=.venv* --statistics
+            black src tests
             pytest -v
