+++ .circleci/config.yml
@@ -17,6 +17,7 @@ jobs:
             python3 -m venv .venv
             . .venv/bin/activate
             pip install -r requirements.txt
+            sudo apt-get install pandoc
       # Step 3: run linter and tests
       - run:
           name: run tests
