+++ src/cli/nbless_click.py
@@ -8,8 +8,7 @@ from nbless.nbless import nbless
 
 
 @click.command()
-@click.argument("in_files", nargs=-1, required=True,
-                type=click.Path(exists=True))
+@click.argument("in_files", nargs=-1, required=True, type=click.Path(exists=True))
 @click.option("-k", "--kernel", "kernel")
 @click.option("-o", "--out_file", "out")
 def nbless_click(in_files: List[str], kernel: str, out: str) -> None:
