+++ pytest.ini
@@ -1,2 +1,0 @@
-[pytest]
-addopts = --mypy --mypy-ignore-missing-imports --doctest-modules
