### 2019-06-14 607f365
rename cli functions and modules and add python_requires to setup.py
