### 2019-06-14 b3bcc0c
add back .travis.yml to try it without tox
