### 2019-06-14 69d5e30
switch travis from pytest to tox
