### 2019-06-14 5292b97
add click to travis
