### 2019-06-14 8f0b0c3
fix click test
