### 2019-06-14 e303b1d
click tests for nbless nbuild and nbexec are working
