+++ tests/test_nbless.py
@@ -30,28 +30,34 @@ def make_temp_notebook(tmp_path: Path) -> str:
     return nb.as_posix()
 
 
-def test_nbuild_cell_contents(tmp_path: Path) -> None:
+def test_nbuild_one_cell(tmp_path: Path) -> None:
     """Run nbuild() to create 3 temporary notebook files from 3 tempfiles."""
     for tempfile in make_tempfiles(tmp_path):
         assert nbuild([tempfile]).cells[0].source == Path(tempfile).read_text()
 
 
-def test_nbless_cell_contents(tmp_path: Path) -> None:
+def test_nbless_one_cell(tmp_path: Path) -> None:
     """Run nbless() to create and execute three notebook files."""
     for tempfile in make_tempfiles(tmp_path):
         assert nbless([tempfile]).cells[0].source == Path(tempfile).read_text()
 
 
-def test_nbuild_cell_type(tmp_path: Path) -> None:
+def test_nbuild_three_cells(tmp_path: Path) -> None:
     """Run nbuild() to create a temporary notebook file from 3 tempfiles."""
-    cells = nbuild(make_tempfiles(tmp_path)).cells
+    files = make_tempfiles(tmp_path)
+    cells = nbuild(files).cells
     assert [c.cell_type for c in cells] == ["markdown", "code", "markdown"]
+    for cell, tempfile in zip(cells, files):
+        assert cell.source == Path(tempfile).read_text()
 
 
-def test_nbless_cell_type(tmp_path: Path) -> None:
+def test_nbless_three_cells(tmp_path: Path) -> None:
     """Run nbless() to create and execute a 3-cell notebook file."""
-    cells = nbless(make_tempfiles(tmp_path)).cells
+    files = make_tempfiles(tmp_path)
+    cells = nbless(files).cells
     assert [c.cell_type for c in cells] == ["markdown", "code", "markdown"]
+    for cell, tempfile in zip(cells, files):
+        assert cell.source == Path(tempfile).read_text()
 
 
 def test_nbexec(tmp_path: Path) -> None:
