+++ tests/test_cli.py
@@ -6,6 +6,7 @@ import pytest
 import click
 from click.testing import CliRunner
 
+from nbless import nbuild
 from cli import nbless_cli, nbuild_cli, nbexec_cli, nbconv_cli
 
 
@@ -25,9 +26,46 @@ def make_tempfiles(tmp_path: Path) -> List[str]:
     return [md.as_posix(), py.as_posix(), txt.as_posix()]
 
 
+def make_temp_notebook(tmp_path: Path) -> str:
+    """Helper function to create a list of pathlib Path objects."""
+    nb = tmp_path / "notebook.ipynb"
+    nb.write_text(nbformat.writes(nbuild(make_tempfiles(tmp_path))))
+    return nb.as_posix()
+
+
 def test_nbless_cli(tmp_path: Path) -> None:
     runner = CliRunner()
     with runner.isolated_filesystem():
-        with open('hello.txt', 'w') as f:
-            f.write('Hello World!')
-        assert Path("hello.txt").read_text() == 'Hello World!'
+        files = make_tempfiles(tmp_path)
+        result = runner.invoke(nbless_cli.nbless_cli, files)
+        cells = nbformat.reads(result.output, as_version=4).cells
+        assert result.exit_code == 0
+        assert [c.cell_type for c in cells] == ["markdown", "code", "markdown"]
+        for cell, tempfile in zip(cells, files):
+            assert cell.source == Path(tempfile).read_text()
+
+
+def test_nbuild_cli(tmp_path: Path) -> None:
+    runner = CliRunner()
+    with runner.isolated_filesystem():
+        files = make_tempfiles(tmp_path)
+        result = runner.invoke(nbuild_cli.nbuild_cli, files)
+        cells = nbformat.reads(result.output, as_version=4).cells
+        assert result.exit_code == 0
+        assert [c.cell_type for c in cells] == ["markdown", "code", "markdown"]
+        for cell, tempfile in zip(cells, files):
+            assert cell.source == Path(tempfile).read_text()
+
+
+def test_nbexec_cli(tmp_path: Path) -> None:
+    """Run nbexec() to execute a temporary notebook file."""
+    runner = CliRunner()
+    with runner.isolated_filesystem():
+        nb = make_temp_notebook(tmp_path)
+        result = runner.invoke(nbexec_cli.nbexec_cli, nb)
+        cells = nbformat.reads(result.output, as_version=4).cells
+        for cell in cells:
+            if cell.cell_type == "code":
+                assert cell.execution_count
+                for output in cell.outputs:
+                    assert output
