### 2019-06-16 bc847ab
Add Gitter badge
