### 2019-06-16 c47d3c6
Update README.rst
