+++ README.rst
@@ -375,7 +375,7 @@ Currently, notebook metadata is lost when using ``nbraze``/``nbuild``/``nbless``
 - Enable ``nbuild``/``nbless`` to accept metadata via a ``metadata.json`` file.
 - Enable ``nbraze`` to output metadata via a ``metadata.json`` file.
 
-.. |Build| image:: https://travis-ci.org/py4ds/nbless.svg?branch=master
+.. |Build| image:: https://travis-ci.org/marskar/nbless.svg?branch=master
    :target: https://travis-ci.org/py4ds/nbless
 .. |Chat| image:: https://badges.gitter.im/py4ds/nbless.svg
    :alt: Join the chat at https://gitter.im/py4ds/nbless
@@ -388,7 +388,7 @@ Currently, notebook metadata is lost when using ``nbraze``/``nbuild``/``nbless``
    :alt: Project Status: Active – The project has reached a stable, usable state and is being actively developed.
    :target: https://www.repostatus.org/#active
 .. |PyUp| image:: https://pyup.io/repos/github/py4ds/nbless/shield.svg
-   :target: https://pyup.io/repos/github/py4ds/nbless/
+   :target: https://pyup.io/repos/github/marskar/nbless/
 .. |Python versions| image:: https://img.shields.io/pypi/pyversions/nbless.svg
    :alt: PyPI - Python Version
    :target: https://www.python.org/downloads/
