### 2019-06-16 e0185d8
Update README.rst
