+++ README.rst
@@ -387,7 +387,7 @@ Currently, notebook metadata is lost when using ``nbraze``/``nbuild``/``nbless``
 .. |Repo status| image:: https://www.repostatus.org/badges/latest/active.svg
    :alt: Project Status: Active – The project has reached a stable, usable state and is being actively developed.
    :target: https://www.repostatus.org/#active
-.. |PyUp| image:: https://pyup.io/repos/github/py4ds/nbless/shield.svg
+.. |PyUp| image:: https://pyup.io/repos/github/marskar/nbless/shield.svg
    :target: https://pyup.io/repos/github/marskar/nbless/
 .. |Python versions| image:: https://img.shields.io/pypi/pyversions/nbless.svg
    :alt: PyPI - Python Version
