### 2019-06-16 3c1038f
nbconv cli test is working
