### 2019-06-17 864bf4f
all test working with 92% coverage
