+++ tests/test_cli.py
@@ -1,35 +1,17 @@
 from pathlib import Path
-from typing import List
 
 import nbformat
-import pytest
 from click.testing import CliRunner
 
-from cli import nbless_cli, nbuild_cli, nbexec_cli, nbconv_cli
-from nbless import nbuild
-
-
-def make_tempfiles(tmp_path: Path) -> List[str]:
-    """Helper function to create a list of pathlib Path objects."""
-    md = tmp_path / "intro.md"
-    py = tmp_path / "plot.py"
-    txt = tmp_path / "discussion.txt"
-    md.write_text("# Introduction\nHere's a plot made with the matplotlib.")
-    py.write_text(
-        "import numpy as np\nimport matplotlib.pyplot as plt\n"
-        "N = 50\nx = y = colors = np.random.rand(N)\n"
-        "area = np.pi * (15 * np.random.rand(N)) ** 2\n"
-        "plt.scatter(x, y, s=area, c=colors, alpha=0.5)\nplt.show()"
-    )
-    txt.write_text("Discussion\nMatplotlib is verbose, but makes cool plots!")
-    return [md.as_posix(), py.as_posix(), txt.as_posix()]
-
-
-def make_temp_notebook(tmp_path: Path) -> str:
-    """Helper function to create a list of pathlib Path objects."""
-    nb = tmp_path / "notebook.ipynb"
-    nb.write_text(nbformat.writes(nbuild(make_tempfiles(tmp_path))))
-    return nb.as_posix()
+from cli import (
+    nbless_cli,
+    nbuild_cli,
+    nbexec_cli,
+    nbconv_cli,
+    nbraze_cli,
+    nbdeck_cli
+)
+from tests.make_temp import make_tempfiles, make_temp_notebook
 
 
 def test_nbless_cli(tmp_path: Path) -> None:
@@ -80,4 +62,30 @@ def test_nbconv_cli(tmp_path: Path) -> None:
         runner.invoke(nbconv_cli.nbconv_cli, [nb, '-o', 'report.html'])
         assert Path('report.html').read_text().startswith('<!DOCTYPE html>\n')
         runner.invoke(nbconv_cli.nbconv_cli, [nb, '-o', 'report.adoc'])
-        assert Path('report.adoc').read_text().startswith('\n[[introduction]]')
+        assert Path('report.adoc').read_text().startswith('\n[[background]]')
+
+
+def test_nbraze_cli(tmp_path: Path):
+    """Extract code and markdown files from the cells of an input notebook."""
+    runner = CliRunner()
+    with runner.isolated_filesystem():
+        runner.invoke(nbraze_cli.nbraze_cli, make_temp_notebook(tmp_path))
+        assert Path("notebook_cell0.md").read_text().startswith("# Background")
+        assert Path("notebook_cell1.py").read_text().startswith("import numpy")
+        assert Path("notebook_cell2.md").read_text().startswith("# Discussion")
+
+
+def test_nbdeck_cli(tmp_path: Path):
+    """ Set up a Jupyter notebook to be viewed as or converted into slides."""
+    runner = CliRunner()
+    with runner.isolated_filesystem():
+        nb = make_temp_notebook(tmp_path)
+        result = runner.invoke(nbdeck_cli.nbdeck_cli, nb)
+        cells = nbformat.reads(result.output, as_version=4).cells
+        c = 0
+        for cell in cells:
+            if cell.cell_type == "markdown" and cell.source.startswith("#"):
+                c += 1
+                assert cell.metadata.slideshow == {"slide_type": "slide"}
+        assert c == 2
+        assert len(cells) == 3
