### 2019-06-17 py4ds/nbless#17 @marskar
I wrote many new tests. I'm at 92% coverage:

```
---------- coverage: platform darwin, python 3.7.3-final-0 -----------
Name                     Stmts   Miss  Cover
--------------------------------------------
src/cli/__init__.py          0      0   100%
src/cli/nbconv_cli.py       10      0   100%
src/cli/nbdeck_cli.py       12      1    92%
src/cli/nbexec_cli.py       13      1    92%
src/cli/nbless_cli.py       14      1    93%
src/cli/nbraze_cli.py       11      0   100%
src/cli/nbuild_cli.py       13      1    92%
src/nbless/__init__.py       9      0   100%
src/nbless/nbconv.py        21      5    76%
src/nbless/nbdeck.py         7      0   100%
src/nbless/nbexec.py         8      0   100%
src/nbless/nbless.py         9      0   100%
src/nbless/nbraze.py        12      3    75%
src/nbless/nbuild.py         8      0   100%
--------------------------------------------
TOTAL                      147     12    92%
```
