### 2019-06-17 3a50434
Bump version to 0.2.27
