+++ docs/source/conf.py
@@ -25,9 +25,9 @@ copyright = '2019, Martin Skarzynski'
 author = 'Martin Skarzynski'
 
 # The short X.Y version
-version = '0.2.18'
+version = '0.2.26'
 # The full version, including alpha/beta/rc tags
-release = '0.2.18'
+release = version
 
 
 # -- General configuration ---------------------------------------------------
