+++ .bumpversion.cfg
@@ -1,0 +1,17 @@
+[bumpversion]
+current_version = 0.2.26
+commit = True
+tag = True
+
+[bumpversion:file:docs/source/conf.py]
+search = version = '{current_version}'
+replace = version = '{new_version}'
+
+[bumpversion:file:setup.py]
+search = version='{current_version}'
+replace = version='{new_version}'
+
+[bumpversion:file:src/nbless/__init__.py]
+search = __version__ = '{current_version}'
+replace = __version__ = '{new_version}'
+
