### 2019-06-17 8a11219
Bump version: 0.2.26 → 0.2.27
