### 2019-06-17 py4ds/nbless#12 @marskar
I use `bumpversion`. I didn't have it setup correctly, but I fixed this by adding a `.bumpversion.cfg` file and editing my `Makefile`.
