### 2019-06-17 421de13
Bump version: 0.2.26 → 0.2.27
