+++ .bumpversion.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 0.2.28
+current_version = 0.2.26
 commit = True
 tag = True
 
