### 2019-06-17 8d00e71
Bump version: 0.2.26 → 0.2.27
