### 2019-06-17 #7 @marskar
- I set up my [versioning system correctly](https://github.com/py4ds/nbless/issues/12)
- I fixed my travis testing problem (turns out the problem was with `tox`)
- I added [many tests](https://github.com/py4ds/nbless/issues/17)
- I made sure the [behaviour of the nbconv matches the docs](https://github.com/py4ds/nbless/issues/14)
The remaining issues all have to do with the docs.
I will try to wrap this up this week.
