### 2019-06-17 4148ec8
working on docs, need to install tests package in virtual env
