+++ docs/source/tests.rst
@@ -1,7 +1,38 @@
-tests
-=====
+tests package
+=============
 
-.. toctree::
-   :maxdepth: 4
+Submodules
+----------
 
-   test_nbless
+tests.make\_temp module
+-----------------------
+
+.. automodule:: tests.make_temp
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+tests.test\_cli module
+----------------------
+
+.. automodule:: tests.test_cli
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+tests.test\_nbless module
+-------------------------
+
+.. automodule:: tests.test_nbless
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+
+Module contents
+---------------
+
+.. automodule:: tests
+   :members:
+   :undoc-members:
+   :show-inheritance:
