### 2019-06-18 600f6a3
Update sphinx-click from 2.0.1 to 2.2.0
