+++ requirements.txt
@@ -14,5 +14,5 @@ tox==3.12.1
 coverage==4.5.3
 Sphinx==2.1.1
 sphinxcontrib.fulltoc==1.2.0
-sphinx-click==2.0.1
+sphinx-click==2.2.0
 twine==1.13.0
