+++ requirements.txt
@@ -4,6 +4,7 @@ click==7.0
 pytest==4.6.3
 pytest-mypy==0.3.2
 pytest-cov==2.7.1
+codecov==2.0.15
 black==19.3b0
 matplotlib==3.1.0
 numpy==1.16.4
