### 2019-06-18 953d1d0
try to add coverage to travis
