### 2019-06-18 fd634b0
try to add coverage to travis
