+++ .travis.yml
@@ -4,7 +4,7 @@ python:
 - 3.6
 - 3.7
 install:
-  - pip install -U pytest-cov pytest-mypy jupyter click matplotlib numpy
+  - pip install -U pytest-cov pytest-mypy jupyter click matplotlib numpy codecov
   - sudo apt-get install pandoc
 script: pytest --verbose --mypy --mypy-ignore-missing-imports --doctest-modules --cov-report term --cov=src/
 deploy:
