### 2019-06-18 5b10f73
remove trailing whitespace in README.rst
