### 2019-06-18 8dcbcf1
add coverage badge
