+++ README.rst
@@ -16,12 +16,12 @@ in
 
 The ``nbless`` Python package consists of 6 Python functions and shell commands:
 
-- ``nbconv``, which converts a notebook into various formats.
-- ``nbdeck``, which prepares a notebook to be viewed as or converted into slides.
-- ``nbexec``, which runs a notebook from top to bottom and saves an executed version.
-- ``nbless``, which calls ``nbuild`` and ``nbexec`` to create and execute a notebook.
-- ``nbraze``, which extracts code and markdown files from a notebook.
-- ``nbuild``, which creates a notebook from source files, e.g. Python (`.py`) and R (`.R`) scripts, markdown (``.md``), and text (``.txt``) files.
+- nbconv_, which converts a notebook into various formats.
+- nbdeck_, which prepares a notebook to be viewed as or converted into slides.
+- nbexec_, which runs a notebook from top to bottom and saves an executed version.
+- nbless_, which calls ``nbuild`` and ``nbexec`` to create and execute a notebook.
+- nbraze_, which extracts code and markdown files from a notebook.
+- nbuild_, which creates a notebook from source files, e.g. Python (`.py`) and R (`.R`) scripts, markdown (``.md``), and text (``.txt``) files.
 
 For a related package that provides programmatic tools for working with `R Markdown <https://rmarkdown.rstudio.com/authoring_quick_tour.html>`__ (Rmd) files,
 check out the `Rmdawn Python package <https://py4ds.github.io/rmdawn/>`__.
@@ -45,62 +45,111 @@ or clone the `repo <https://github.com/py4ds/nbless>`__, e.g.
 using setup.py (``python setup.py install``) or ``pip``
 (``pip install .``).
 
-R and Python interoperability
------------------------------
+Usage
+-----
 
-All of the Nbless functions and commands work for Python *and* R code, with the caveat
-that ``nbexec`` and ``nbless`` require the kernel argument to be set to
-``ir`` if R code files (``.R``) are included.
+.. _nbconv:
 
-If you want to execute a Jupyter notebook that contains both R and
-Python code, you will have to put the R code in Python scripts (``.py``)
-and use the `rpy2 <https://rpy2.readthedocs.io/>`__ Python library with the default kernel
-(``python3``).
+Converting Jupyter notebooks with ``nbconv``
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-Code files extracted from code cells with ``nbraze`` must all have the same extension (e.g. `.py`).
-The ``language_info`` key is not always defined, so ``nbraze`` does not try to infer the programming language from Jupyter notebook metadata.
+The ``nbconv`` shell command and Python function can export a
+notebook to many different formats using the ``nbconvert`` library.
+Converting to all formats except HTML requires pandoc.
+Exporting to PDF requires LaTeX.
 
-You can also run the Nbless functions in an R environment using the
-`reticulate <https://rstudio.github.io/reticulate/>`__ R package.
+The supported exporters are
 
-All Nbless functions and commands rely on the `nbconvert <https://nbconvert.readthedocs.io/>`__ and `nbformat <http://nbformat.readthedocs.io/>`__ modules that are included with the ``jupyter`` Python library.
[…]
