+++ README.rst
@@ -53,7 +53,7 @@ Usage
 Converting Jupyter notebooks with ``nbconv``
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-The ``nbconv`` shell command and Python function can export a
+The ``nbconv`` shell command can export a
 notebook to many different formats using the ``nbconvert`` library.
 Converting to all formats except HTML requires pandoc.
 Exporting to PDF requires LaTeX.
@@ -100,132 +100,209 @@ If neither the ``exporter`` or ``out_file`` arguments are provided, the exporter
 
     nbconv notebook.ipynb
 
+Unlike the shell command,
+the ``nbconv`` Python function does not create a file on its own.
+To create a converted file with Python, use the ``pathlib`` library.
+
 .. code:: python
 
     from pathlib import Path
-    import nbformat
     from nbless import nbconv
 
-    # Create notebook.py from notebook.ipynb in notebooks folder
-    # nbconv() returns a filename and file contents as strings
-    def write_file(filename: str, contents: str) -> None:
-        with open(filename, 'w') as f:
-            f.write(contents)
+    # Create notebook.py from notebook.ipynb
+    filename, contents = nbconv("notebook.ipynb", "python")
+    Path(filename).write_text(contents)
 
-    filename, contents = nbconv("notebooks/notebook.ipynb")
-    write_file(filename, contents)
-    write_file(*nbconv("notebooks/notebook.ipynb"))
-
-    # Create notebook.html from notebook.ipynb in notebooks folder
-    write_file(*nbconv("notebooks/notebook.ipynb", "html"))
-
-    # Create script.py from notebook.ipynb in notebooks folder
-    write_file('script.py', nbconv("notebooks/notebook.ipynb")[1])
-
-    # Create report.html from notebook.ipynb in notebooks folder
-    write_file('report.html', nbconv("notebooks/notebook.ipynb", 'html')[1])
-
-    # Create HTML slides from notebook.ipynb in notebooks folder
-    # nbdeck() returns a filename and file contents as strings
-    nbformat.write(nbdeck("notebook.ipynb"), "slides.ipynb", version=4)
-    filename, contents = nbconv("slides.ipynb", "slides")
-    write_file(filename, contents)
-    write_file(*nbconv("notebooks/notebook.ipynb", "slides"))
-
+    # Create report.html from notebook.ipynb
+    filename, contents = nbconv("notebook.ipynb", "html")
+    Path('report.html').write_text(contents)
 
 .. _nbdeck:
 
 Creating HTML slides with ``nbdeck`` and ``nbconv``
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-With ``nbdeck``, you can prepare Jupyter slides from source files (e.g. ``source_file1.md``, ``source_file2.py``) like this:
+With ``nbdeck``, you can prepare HTML slides from a Jupyter notebook.
+
 
 .. code:: sh
 
-    nbless slide_file* -o slides.ipynb
-    nbdeck slides.ipynb -o slides.ipynb
+    nbdeck notebook.ipynb -o slides.ipynb
     nbconv slides.ipynb  -e slides -o slides.html
+
+You can run ``nbdeck`` without ``nbconv``,
+if you do not want to create HTML slides and instead want to use
+`nbviewer <https://nbviewer.jupyter.org/>`__ or the
+`RISE extension <https://github.com/damianavila/RISE#rise>`__.
+If an ``out_file`` name is not provided, the notebook file contents will be
+printed.
+You can provide a more descriptive name for the executed output notebook with
[…]
