+++ README.rst
@@ -1,7 +1,6 @@
 Nbless: a Python package for programmatic Jupyter notebook workflows
 ====================================================================
 
-
 |Build| |Chat| |Coverage| |License| |PyPI| |Python versions| |PyUp| |Repo status|
 
 Introduction
@@ -123,7 +122,6 @@ Creating HTML slides with ``nbdeck`` and ``nbconv``
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 With ``nbdeck``, you can prepare HTML slides from a Jupyter notebook.
-
 
 .. code:: sh
 
@@ -262,7 +260,6 @@ To create a notebook file, use the ``nbformat`` library.
 	Rnb = nbless(["plot.R", "notes.txt"], kernel="ir")
     nbformat.write(Rnb, "Rexecuted.ipynb")
 
-
 .. _nbraze:
 
 Extracting source files from a Jupyter notebook with ``nbraze``
@@ -312,7 +309,6 @@ cells <https://jupyter-notebook.readthedocs.io/en/stable/examples/Notebook/Worki
 
     nbuild README.md plot.py notes.txt > notebooks/notebook.ipynb
 
-
 Instead of redirecting to a file (``>``), you can use the ``--out_file``
 or ``-o`` flag:
 
@@ -339,60 +335,6 @@ To create a notebook file, use the ``nbformat`` library.
     # Create notebook.ipynb from plot.py and notes.txt
     nb = nbuild(["plot.py", "notes.txt"])
     nbformat.write(nb, "notebook.ipynb")
-
-
-Missing a dependency?
-~~~~~~~~~~~~~~~~~~~~~
-
-If you installed via ``pip`` or ``setup.py``, you should have both of
-the dependencies (``click`` and ``jupyter``) already. If not, try pip
-installing them separately.
-
-.. code:: sh
-
-    pip install jupyter click
-
-If you have `Anaconda <https://www.anaconda.com/download/>`__ or
-`Miniconda <https://conda.io/miniconda.html>`__ installed, you can run
-
-.. code:: sh
-
-    conda install -yc conda-forge jupyter click
-
-Too many file names to type out?
---------------------------------
-
-The easiest way to handle large numbers of files is to use the ``*`` wildcard in the shell.
-
-.. code:: sh
-
-    nbuild source_file* -o notebook.ipynb
-
-You can use the ``ls`` command to assign all of the relevant names in
-the current directory to a variable and pass this variable as an
-argument.
-
-Consider the example below:
-
-.. code:: sh
-
-    touch {01..09}.py
-    name_list=`ls 0*.py`
-    nbuild `echo $name_list`
-
-In Python environments, ``os.listdir`` can provide a list of
-all files:
-
-.. code:: python
-
-    from os import listdir
-    from os.path import isfile, join
-    onlyfiles = [f for f in listdir(mypath) if isfile(join(mypath, f))]
-
-To preserve the order and differentiate files that should be
-incorporated into the notebook, it may be helpful to left pad file names
-with zeros (e.g. ``01\_intro.md``, ``02\_figure1.R``).
-This works well for R scripts, but Python files that start with numbers cannot be imported.
 
 Related projects
 ----------------
