### 2019-06-18 py4ds/nbless#16 @marskar
The docs are built (in-part) from the README.
I simplified and streamlined the docs.
I created a few new sections rather than including everything in the docs.
If you have any suggestions of how to further split things up, please let me know.
