### 2019-06-18 py4ds/nbless#16 @ocefpaf
> If you have any suggestions of how to further split things up, please let me know.

I'll take a second look soon but it looks good to me the way it is.
