### 2019-06-19 d63027b
improve tests
