+++ tests/test_nbless.py
@@ -1,26 +1,27 @@
 from pathlib import Path
 
+import nbformat
 import pytest
-from tests.make_temp import make_tempfiles, make_temp_notebook
+from tests.make_temp import make_files, make_notebook, exec_notebook
 
 from nbless import nbless, nbuild, nbexec, nbconv, nbraze, nbdeck
 
 
 def test_nbuild_one_cell(tmp_path: Path) -> None:
     """Run nbuild() to create 3 temporary notebook files from 3 tempfiles."""
-    for tempfile in make_tempfiles(tmp_path):
+    for tempfile in make_files(tmp_path):
         assert nbuild([tempfile]).cells[0].source == Path(tempfile).read_text()
 
 
 def test_nbless_one_cell(tmp_path: Path) -> None:
     """Run nbless() to create and execute three notebook files."""
-    for tempfile in make_tempfiles(tmp_path):
+    for tempfile in make_files(tmp_path):
         assert nbless([tempfile]).cells[0].source == Path(tempfile).read_text()
 
 
 def test_nbuild_three_cells(tmp_path: Path) -> None:
     """Run nbuild() to create a temporary notebook file from 3 tempfiles."""
-    files = make_tempfiles(tmp_path)
+    files = make_files(tmp_path)
     cells = nbuild(files).cells
     assert [c.cell_type for c in cells] == ["markdown", "code", "markdown"]
     for cell, tempfile in zip(cells, files):
@@ -29,7 +30,7 @@ def test_nbuild_three_cells(tmp_path: Path) -> None:
 
 def test_nbless_three_cells(tmp_path: Path) -> None:
     """Run nbless() to create and execute a 3-cell notebook file."""
-    files = make_tempfiles(tmp_path)
+    files = make_files(tmp_path)
     cells = nbless(files).cells
     assert [c.cell_type for c in cells] == ["markdown", "code", "markdown"]
     for cell, tempfile in zip(cells, files):
@@ -38,7 +39,7 @@ def test_nbless_three_cells(tmp_path: Path) -> None:
 
 def test_nbexec(tmp_path: Path) -> None:
     """Run nbexec() to execute a temporary notebook file."""
-    for cell in nbexec(make_temp_notebook(tmp_path)).cells:
+    for cell in nbexec(make_notebook(tmp_path)).cells:
         if cell.cell_type == "code":
             assert cell.execution_count
             for output in cell.outputs:
@@ -48,21 +49,43 @@ def test_nbexec(tmp_path: Path) -> None:
 @pytest.mark.parametrize("not_exporters", ["htm", "ipython", "markup"])
 def test_raises(not_exporters, tmp_path: Path) -> None:
     """Make sure a ValueError is raised if nbconv() gets a bad exporter."""
-    nb = make_temp_notebook(tmp_path)
+    nb = make_notebook(tmp_path)
     with pytest.raises(ValueError):
         nbconv(in_file=nb, exporter=not_exporters)
+        nbconv(in_file=nb, out="out." + not_exporters)
 
 
 @pytest.mark.parametrize("exporters", ["html", "asciidoc", "rst"])
 def test_nbconv(exporters, tmp_path: Path) -> None:
     """Convert ``tempfiles`` with each exporter in ``exporters``."""
-    nb = make_temp_notebook(tmp_path)
+    nb = make_notebook(tmp_path)
     assert nbconv(in_file=nb, exporter=exporters)[0].endswith("." + exporters)
+    assert nbconv(in_file=nb)[0].endswith(".html")
+
+
+def test_nbconv_file_contents(tmp_path: Path):
+    nb = make_notebook(tmp_path)
[…]
