### 2019-06-19 53cc70f
improve tests up to 98% coverage
