### 2019-06-19 py4ds/nbless#15 @marskar
I removed this section as part of the effort to simplify the docs. The most important parts are with the code for each shell command.
