### 2019-06-19 py4ds/nbless#13 @marskar
`nbexec` used to work like that, but I decided this was inconsistent with the other commands. I updated the docs to match the current code. Thanks for catching this!
