### 2019-06-19 #7 @ocefpaf
@lwasser I believe that @marskar addressed all the issues I raised in my review. (I updated the check boxes in https://github.com/pyOpenSci/software-review/issues/7#issuecomment-498061890 to reflect that.)
