### 2019-06-19 #7 @lwasser
awesome. ok so we just need a second review now i think? we ran into an issue where we need reviewers for earthpy too.. and so i think @leouieda was asked twice which isn't ideal. Leo would you prefer to review this package or earthpy? we can try to find others for earthpy if need be!

and thank you both for working through this process - it seemed pretty seamless the way you implemented the review and changes.
