### 2019-06-19 8480c66
rename out argument to out_file and work on tests
