+++ src/nbless/nbconv.py
@@ -5,11 +5,11 @@ from typing import Tuple
 from nbconvert.exporters.base import get_exporter
 
 
-def nbconv(in_file: str, exporter: str = "", out: str = "") -> Tuple[str, str]:
+def nbconv(in_file: str, exporter: str = "", out_file: str = "") -> Tuple[str, str]:
     """Convert a notebook into various formats using ``nbformat`` exporters.
 
     :param in_file: The name of the input notebook file.
-    :param out: The name of the output file.
+    :param out_file: The name of the output file.
     :param exporter: The exporter that determines the output file type.
     :return: A 2-tuple of the output file's 1) name and 2) contents.
     :note: The exporter type must be 'asciidoc', 'pdf', 'html', 'latex',
@@ -18,11 +18,11 @@ def nbconv(in_file: str, exporter: str = "", out: str = "") -> Tuple[str, str]:
            Exporting to pdf requires latex.
     """
     in_path = Path(in_file)
-    if not out and not exporter:
+    if not out_file and not exporter:
         print("Setting exporter to HTML!")
         exporter = "html"
-    elif out and not exporter:
-        out_path = Path(out)
+    elif out_file and not exporter:
+        out_path = Path(out_file)
         ext_exp_dict = {
             ".asciidoc": "asciidoc",
             ".adoc": "asciidoc",
@@ -42,8 +42,8 @@ def nbconv(in_file: str, exporter: str = "", out: str = "") -> Tuple[str, str]:
             print("Setting exporter to HTML!")
             exporter = "html"
     contents, resources = get_exporter(exporter)().from_filename(in_file)
-    if not out:
+    if not out_file:
         name = in_path.stem + resources.get("output_extension", ".txt")
         return name, contents
     else:
-        return out, contents
+        return out_file, contents
