### 2019-06-19 72230c3
improve tests up to 99% coverage
