+++ docs/_sources/nbless.rst.txt
@@ -1,10 +1,0 @@
-nbless package
-==============
-
-Module contents
----------------
-
-.. automodule:: nbless
-   :members:
-   :undoc-members:
-   :show-inheritance:
