+++ docs/_sources/index.rst.txt
@@ -1,22 +1,0 @@
-Welcome to Nbless's documentation!
-======================================================
-
-.. toctree::
-   :maxdepth: 2
-   :caption: Contents:
-
-   Project overview <readme.rst>
-   Command-line interface <cli.rst>
-   Module reference <modules.rst>
-   Test reference <test_modules.rst>
-   R and Python interoperability <interop.rst>
-   Missing dependencies? <dep.rst>
-   Too many files to type out? <many-file.rst>
-
-
-Indices and tables
-==================
-
-* :ref:`genindex`
-* :ref:`modindex`
-* :ref:`search`
