### 2019-06-19 0ef5cc6
fix .bumpversion.cfg
