+++ .bumpversion.cfg
@@ -4,14 +4,14 @@ commit = True
 tag = True
 
 [bumpversion:file:docs/source/conf.py]
-search = version = '{current_version}'
-replace = version = '{new_version}'
+search = '{current_version}'
+replace = '{new_version}'
 
 [bumpversion:file:setup.py]
-search = version='{current_version}'
-replace = version='{new_version}'
+search = '{current_version}'
+replace = '{new_version}'
 
 [bumpversion:file:src/nbless/__init__.py]
-search = __version__ = '{current_version}'
-replace = __version__ = '{new_version}'
+search = '{current_version}'
+replace = '{new_version}'
 
