+++ .bumpversion.cfg
@@ -4,14 +4,14 @@ commit = True
 tag = True
 
 [bumpversion:file:docs/source/conf.py]
-search = "{current_version}"
-replace = "{new_version}"
+search = {current_version}
+replace = {new_version}
 
 [bumpversion:file:setup.py]
-search = "{current_version}"
-replace = "{new_version}"
+search = {current_version}
+replace = {new_version}
 
 [bumpversion:file:src/nbless/__init__.py]
-search = "{current_version}"
-replace = "{new_version}"
+search = {current_version}
+replace = {new_version}
 
