+++ docs/_sources/cli.rst.txt
@@ -1,0 +1,20 @@
+Nbless command-line interface (CLI)
+===================================
+
+.. click:: cli.nbless_cli:nbless_cli
+  :prog: nbless
+
+.. click:: cli.nbraze_cli:nbraze_cli
+  :prog: nbraze
+
+.. click:: cli.nbdeck_cli:nbdeck_cli
+  :prog: nbdeck
+
+.. click:: cli.nbuild_cli:nbuild_cli
+  :prog: nbuild
+
+.. click:: cli.nbconv_cli:nbconv_cli
+  :prog: nbconv
+
+.. click:: cli.nbexec_cli:nbexec_cli
+  :prog: nbexec
