### 2019-06-19 #7 @leouieda
Hi all, sorry I haven't been able to stay up to date with my github notifications. I don't know if I'll be able to get this done until the end of month.

@lwasser I don't have a preference but I feel like it would be easier to find someone else to review nbless than earthpy, since nbless is general purpose. I could do earthpy if you prefer.
