### 2019-06-20 2392755
improve tests up to 100% coverage
