+++ tests/test_nbless.py
@@ -8,19 +8,19 @@ from nbless import nbless, nbuild, nbexec, nbconv, nbraze, nbdeck
 
 
 def test_nbuild_one_cell(tmp_path: Path) -> None:
-    """Run nbuild() to create 3 temporary notebook files from 3 tempfiles."""
+    """Run ``nbuild`` to create 3 temporary notebook files from 3 tempfiles."""
     for tempfile in make_files(tmp_path):
         assert nbuild([tempfile]).cells[0].source == Path(tempfile).read_text()
 
 
 def test_nbless_one_cell(tmp_path: Path) -> None:
-    """Run nbless() to create and execute three notebook files."""
+    """Run ``nbless`` to create and execute three notebook files."""
     for tempfile in make_files(tmp_path):
         assert nbless([tempfile]).cells[0].source == Path(tempfile).read_text()
 
 
 def test_nbuild_three_cells(tmp_path: Path) -> None:
-    """Run nbuild() to create a temporary notebook file from 3 tempfiles."""
+    """Run ``nbuild`` to create a temporary notebook file from 3 tempfiles."""
     files = make_files(tmp_path)
     cells = nbuild(files).cells
     assert [c.cell_type for c in cells] == ["markdown", "code", "markdown"]
@@ -29,7 +29,7 @@ def test_nbuild_three_cells(tmp_path: Path) -> None:
 
 
 def test_nbless_three_cells(tmp_path: Path) -> None:
-    """Run nbless() to create and execute a 3-cell notebook file."""
+    """Run ``nbless`` to create and execute a 3-cell notebook file."""
     files = make_files(tmp_path)
     cells = nbless(files).cells
     assert [c.cell_type for c in cells] == ["markdown", "code", "markdown"]
@@ -38,7 +38,7 @@ def test_nbless_three_cells(tmp_path: Path) -> None:
 
 
 def test_nbexec(tmp_path: Path) -> None:
-    """Run nbexec() to execute a temporary notebook file."""
+    """Run ``nbexec`` to execute a temporary notebook file."""
     for cell in nbexec(make_notebook(tmp_path)).cells:
         if cell.cell_type == "code":
             assert cell.execution_count
@@ -48,7 +48,7 @@ def test_nbexec(tmp_path: Path) -> None:
 
 @pytest.mark.parametrize("not_exporters", ["htm", "ipython", "markup"])
 def test_raises(not_exporters, tmp_path: Path) -> None:
-    """Make sure a ValueError is raised if nbconv() gets a bad exporter."""
+    """Make sure a ``ValueError`` is raised if ``nbconv`` gets a bad exporter."""
     nb = make_notebook(tmp_path)
     with pytest.raises(ValueError):
         nbconv(in_file=nb, exporter=not_exporters)
@@ -57,18 +57,21 @@ def test_raises(not_exporters, tmp_path: Path) -> None:
 
 @pytest.mark.parametrize("exporters", ["html", "asciidoc", "rst"])
 def test_nbconv(exporters, tmp_path: Path) -> None:
-    """Convert ``tempfiles`` with each exporter in ``exporters``."""
+    """Convert a temporary notebook with each exporter in ``exporters``."""
     nb = make_notebook(tmp_path)
     assert nbconv(in_file=nb, exporter=exporters)[0].endswith("." + exporters)
     assert nbconv(in_file=nb)[0].endswith(".html")
 
 
 def test_nbconv_file_contents(tmp_path: Path):
+    """Run ``nbconv`` with the ``exporter`` or ``out_file`` argument."""
     nb = make_notebook(tmp_path)
[…]
