### 2019-06-20 91c65bc
remove coverage from .travis.yml
