### 2019-06-20 #7 @marskar
Hi @lwasser, @ocefpaf, and @leouieda,
Indeed, the review was seamless and very helpful.
No rush with the second review.
I made many improvements including getting the test coverage up to 100%.
I would like to volunteer as a reviewer for earthpy if you think that would work.
We can discuss this on the call tomorrow.
