### 2019-06-20 #7 @lwasser
@marskar that would be wonderful. it sounds like you and @leouieda could do earthpy and then maybe @kysolvik or @carsonfarmer could be interested in reviewing nbless. i'll put your names in the doc and we can chat today. @luizirber note that we may have two reviewers for earthpy now!!
