### 2019-06-20 #7 @lwasser
@kysolvik are you interested in reviewing this package? Leo is open to reviewing earthpy. please let me know!!
