+++ requirements.txt
@@ -11,7 +11,7 @@ numpy==1.16.4
 pypandoc==1.4
 bumpversion==0.5.3
 wheel==0.33.4
-tox==3.12.1
+tox==3.13.1
 coverage==4.5.3
 Sphinx==2.1.2
 sphinxcontrib.fulltoc==1.2.0
