+++ requirements.txt
@@ -2,7 +2,7 @@
 jupyter==1.0.0
 click==7.0
 pytest==4.6.3
-pytest-mypy==0.3.2
+pytest-mypy==0.3.3
 pytest-cov==2.7.1
 codecov==2.0.15
 black==19.3b0
