+++ requirements.txt
@@ -1,7 +1,7 @@
 -e .
 jupyter==1.0.0
 click==7.0
-pytest==4.6.3
+pytest==5.0.0
 pytest-mypy==0.3.2
 pytest-cov==2.7.1
 codecov==2.0.15
