### 2019-07-10 py4ds/nbless#30 @kysolvik
Docs sidebar doesn't scroll when it fills up
On smaller screens, not all of the subheadings appear in the sidebar of the docs site. Sidebar isn't scrollable. Pretty minor issue, but makes navigation a little harder.

Screenshot from my 12.5" screen:
![docs_sidebar_scroll](https://user-images.githubusercontent.com/24379590/61011211-dc3c1f80-a336-11e9-93f4-d81fef0df968.png)

xref.: pyOpenSci/software-review#7
