### 2019-07-11 py4ds/nbless#13 @kysolvik
It looks like there still might be some leftovers in the docs from when the behavior was different. These lines in the readme make it sound like it'll create a output notebook by default.

"You can provide a more descriptive name for the executed output notebook..."

and

"You can preview the default output filename and the raw notebook output by running nbexec with only the positional argument:"
