### 2019-07-11 py4ds/nbless#31 @kysolvik
No contribution guidelines
I couldn't find any contribution or community guidelines. If they're not there, could be worth adding a couple lines to the readme.

xref.: pyOpenSci/software-review#7
