### 2019-07-12 #7 @kysolvik
*KS*: Maybe consider moving the nbuild vignette before the nbless one? I was unclear on how nbless would build the notebook, but the nbuild section explains it.

- [X] **Function Documentation:** for all user-facing functions
- [X] **Examples** for all user-facing functions
- [ ] **Community guidelines** including contribution guidelines in the README or CONTRIBUTING, and DESCRIPTION with `URL`, `BugReports` and `Maintainer`.

*KS*: I didn't see these.

**Readme requirements**

re
- [X] Badges for continuous integration and test coverage, the badge for pyOpenSci peer-review once it has started (see below), a repostatus.org badge, and any other badges. If the README has many more badges, you might want to consider using a table for badges, see this example, that one and that one. Such a table should be more wide than high.
- [X] Short description of goals of package, with descriptive links to all vignettes (rendered, i.e. readable, cf the documentation website section) unless the package is small and there’s only one vignette repeating the README.
- [X] Any additional setup required (authentication tokens, etc)
- [X] Brief demonstration usage
- [X] Direction to more detailed documentation (e.g. your documentation files or website).
- [ ] If applicable, how the package compares to other similar packages and/or how it relates to other packages

*KS*: In the related projects section, could include a blurb about how nbless is different.

*KS: General review notes on docs*
Minor issue with the docs sidebar not being scrollable (py4ds/nbless#30).

I think the README is still a bit lengthy and dense, but it's not a big deal if it's hard to break up. Ideally
 maybe the README could just have the nbconv section and then a list of the rest of the functionality with links to the docs for usage examples. But I totally leave this to you if you want to have all the commands in the README.

For nbexec, still some inconsistencies in docs (py4ds/nbless#13).

- [X] **Continuous Integration:** Has continuous integration, such as Travis CI, AppVeyor, CircleCI, and/or others.

---

Awesome package and I think it's a great fit for pyOpenSci. The CLI is super easy to use. I can imagine it being used frequently, especially for education. I opened a couple issues and added a couple comments, but they're all minor. I opened a few issues. Let me know if I can help at all!

**Summary of the issues:**
- Sidebar in docs doesn't scroll (py4ds/nbless#30)
- Still some docs inconsistencies for nbexec (py4ds/nbless#13)
- Add contribution guidelines (could just be a line or two in the README) (py4ds/nbless#31)
- Add note in README or docs about how nbless is different from the related packages (py4ds/nbless#32)
- Consider moving nbuild vignette to above the nbless one for clarity. (Didn't create an issue, just a suggestion).
