### 2019-07-12 py4ds/nbless#32 @kysolvik
Add detail on what makes nbless different from related packages (e.g. nbconvert, jupytest)
Could just be a short blurb in the Related Projects section of the readme or whatever you think is best!

xref.: pyOpenSci/software-review#7
