### 2019-07-15 #7 @lwasser
thank you @kysolvik !! @marskar it looks like you saw the review. please let us know if you have any questions or comments. when you are done addressing Kylen's comments above, please ping me!!
