### 2019-07-16 py4ds/nbless#30 @marskar
Hi @kysolvik,
I tried to address this problem in docs/source/conf.py with this commit marskar@b665916.
