### 2019-07-16 py4ds/nbless#31 @marskar
Hi @kysolvik,
I added a `CONTRIBUTING.md file` based on [this gist](https://gist.github.com/PurpleBooth/b24679402957c63ec426) with this commit 00d3223. I think that this is better than including this information in the README. Please let me know if you disagree.
