### 2019-07-16 dd2a5eb
explain what differentiates nbless from other packages
