### 2019-07-16 py4ds/nbless#32 @marskar
Great suggestion! I added a few lines about what differentiates `nbless` from other packages with this commit 00d3223.
