### 2019-07-16 53192da
update docs to match new behavior of nbexec and nbconv
