### 2019-07-16 py4ds/nbless#13 @marskar
Hi @kysolvik,
You're right, I closed this issue to quickly! I made new changes with this commit dd2a5eb.
I think the docs are correct now.
