### 2019-07-16 #7 @lwasser
wow! so @kysolvik @ocefpaf it looks like @marskar has completed everything requested in this review. do you guys want to have one last look at things and then let me know if we are good to ACCEPT this package as our first in the pyopensci ecosystem. BTW this means we will need to create a badge for it!! :)
