### 2019-07-16 #7 @kysolvik
Looks good to me @marskar and @lwasser!! Good work all, especially @marskar for creating an awesome package!
