### 2019-07-17 45ba05b
edit nbraze test to recognize metadata file
