### 2019-07-17 b1bc9a9
edit language info test to recognize metadata file
