### 2019-07-17 #7 @lwasser
yahoo!! ok i'll followup with a badge. i don't think anyone has respond to my discourse post yet. and i'll double check we aren't missing anything else that ROS does when accepting a package!
