### 2019-07-22 bd48552
add pyOpenSci badge to readme, woot!
