### 2019-07-22 8b643b6
update readme and docs with pyOpenSci badge plus logo
