### 2019-07-26 2b22bbb
Update numpy from 1.16.4 to 1.17.0
