+++ requirements.txt
@@ -7,7 +7,7 @@ pytest-cov==2.7.1
 codecov==2.0.15
 black==19.3b0
 matplotlib==3.1.1
-numpy==1.16.4
+numpy==1.17.0
 pypandoc==1.4
 bumpversion==0.5.3
 wheel==0.33.4
