+++ requirements.txt
@@ -12,7 +12,7 @@ pypandoc==1.4
 bumpversion==0.5.3
 wheel==0.33.4
 tox==3.13.2
-coverage==4.5.3
+coverage==4.5.4
 Sphinx==2.1.2
 sphinxcontrib.fulltoc==1.2.0
 sphinx-click==2.2.0
