### 2019-08-06 #7 @lwasser
ok this package has been ACCEPTED into pyopensci. it lives here: https://github.com/marskar/nbless/ . we now need to put it somewhere on our website so there is a connection between it and pyopensci. i'll create a separate issue for that  in our website repo and ill close this here!!

https://github.com/pyOpenSci/pyopensci.github.io/issues/3
