### 2022-10-19 #7 @lwasser
hey @marskar this is related to both the survey ([which i'd greatly appreciate your filling out ](https://forms.gle/DF4qoAdXsCUuerj48) But also I noticed that this package hasn't been updated since 2019. I'm just wondering if this package is still maintained by you or someone? Of if we need to sunset / archive it . it has some stars but i haven't look at the stats to see if people are using it. Please reply here and let me know. it's important that packages in our ecosystem are maintained. And if they aren't we need to either find a new reviewer OR we need to help you sunset and archive it so people know it's not actively maintained.

Many thanks for responding. you can also email me if that is easier at leah at pyopensci.org
All the best.
