### 2024-12-15 #7 @banesullivan
I've add the [archived label](https://github.com/pyOpenSci/software-submission/issues?q=label%3Aarchived) to this review

cc @lwasser
