### 2024-12-17 #7 @lwasser
oh this reminds me @banesullivan we have a [workflow](https://github.com/pyOpenSci/software-submission/blob/main/.github/workflows/update-project.yml) that moves packages on our project board based on the label setting. we will want to add archive https://github.com/pyOpenSci/software-submission/blob/main/update-project.json to the json file so archived packages are moved too!

i started a [doc to document this process](https://hackmd.io/@lawasser/SJzCuVkHJg). let's use this to document this and also the label workflow!!
